Use GCC <4.3 semantics for inlining to avoid linker failures
C99: Section 6.7.4p6 of the C99 Standard states the following: "An inline
definition does not provide an external definition for the function, and does
not forbid an external definition in another translation unit. An inline
definition provides an alternative to an external definition, which a translator
may use to implement any call to the function in the same translation unit. It
is unspecified whether a call to the function uses the inline definition or the
external definition."

This may cause linker failures when functions are marked inline or __inline.
Using -fgnu89-inline avoids this behaviour.

Fixes: #3